+++ src/js/admin/acf-options-page/index.js
@@ -1,42 +1,26 @@
 import $ from 'jquery';
 
 $(document)
-.on('click', '.acf-wpo-import-export [data-action="export"][data-ajax-request]', e => {
+	.on('change', '#acf-wpo-import-file', e => {
+		const fileInput = e.target;
+		let fileReader;
+		if ( !! fileInput.value ) {
 
-	e.preventDefault();
 
-	const request_data = JSON.parse( $(e.target).attr('data-ajax-request') );
-	const $textarea = $('.acf-wpo-import-export textarea')
+			fileReader = new FileReader()
+			fileReader.onload = () => {
+				$('#acf-wpo-import-json').val( fileReader.result )
+				$('#import').prop('disabled', false );
+			}
 
-	$.ajax( {
-		url: wp.ajax.settings.url,
-		data: request_data,
-		method:'post',
-		success: response => {
-			$textarea.val(JSON.stringify(response));
-		}
-	} );
-
-})
-.on('click', '.acf-wpo-import-export [data-action="import"][data-ajax-request]', e => {
-
-	e.preventDefault();
-
-	const request_data = JSON.parse( $(e.target).attr('data-ajax-request') );
-	const $textarea = $('.acf-wpo-import-export textarea')
-	request_data.data = JSON.parse( $textarea.val() );
-	if ( !! request_data.data )
-	$.ajax( {
-		url: wp.ajax.settings.url,
-		method:'post',
-		data: request_data,
-		success: response => {
-			if ( response.success ) {
-				location.reload()
-			} else {
-				console.log(response.message)
+			if ( !! fileInput.files[0] ) {
+				fileReader.readAsText( fileInput.files[0] )
 			}
 		}
-	} );
 
-})
+	})
+	.on('click', '.acf-wpo-export [type="submit"]', e => {
+
+		setTimeout( () => acf.unlockForm( $(e.target).closest('form') ) );
+
+	})